python-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
torch==2.3.0
torchvision==0.18.0
scikit-image==0.23.2
scipy==1.13.0
httpx==0.27.0
pydantic==2.7.0
# YOLOv8 (auto-downloads yolov8n.pt and yolov8n-seg.pt on first use)
ultralytics==8.2.0
# Detectron2 — install separately, see MODELS.md
# pip install 'git+https://github.com/facebookresearch/detectron2.git'
